updating test lint
