+++ koku/masu/processor/aws/aws_report_processor.py
@@ -22,12 +22,9 @@ import json
 import logging
 from os import path
 
-from tenant_schemas.utils import schema_context
-
 from masu.config import Config
 from masu.database import AWS_CUR_TABLE_MAP
 from masu.database.aws_report_db_accessor import AWSReportDBAccessor
-from masu.database.report_manifest_db_accessor import ReportManifestDBAccessor
 from masu.database.reporting_common_db_accessor import ReportingCommonDBAccessor
 from masu.processor.report_processor_base import ReportProcessorBase
 from reporting.provider.aws.models import (AWSCostEntry,
@@ -86,6 +83,7 @@ class AWSReportProcessor(ReportProcessorBase):
             report_path=report_path,
             compression=compression,
             provider_id=provider_id,
+            manifest_id=manifest_id,
             processed_report=ProcessedReport()
         )
 
@@ -119,7 +117,7 @@ class AWSReportProcessor(ReportProcessorBase):
 
         """
         row_count = 0
-        self._delete_line_items()
+        self._delete_line_items(AWSReportDBAccessor, self.column_map)
         opener, mode = self._get_file_opener(self._compression)
         is_finalized_data = self._check_for_finalized_bill()
         # pylint: disable=invalid-name
@@ -171,32 +169,6 @@ class AWSReportProcessor(ReportProcessorBase):
             row = reader.__next__()
             invoice_id = row.get('bill/InvoiceId')
             return invoice_id is not None and invoice_id != ''
-
-    def _delete_line_items(self):
-        """Delete stale data for the report being processed, if necessary."""
-        if not self.manifest_id:
-            return False
-
-        with ReportManifestDBAccessor() as manifest_accessor:
-            manifest = manifest_accessor.get_manifest_by_id(self.manifest_id)
-            if manifest.num_processed_files != 0:
-                return False
-            # Override the bill date to correspond with the manifest
-            bill_date = manifest.billing_period_start_datetime.date()
-            provider_id = manifest.provider_id
-
-        LOG.info('Deleting data for schema: %s and bill date: %s',
-                 self._schema_name, str(bill_date))
-
-        with AWSReportDBAccessor(self._schema_name, self.column_map) as accessor:
-            bills = accessor.get_cost_entry_bills_query_by_provider(provider_id)
-            bills = bills.filter(billing_period_start=bill_date).all()
-            with schema_context(self._schema_name):
-                for bill in bills:
-                    line_item_query = accessor.get_lineitem_query_for_billid(bill.id)
-                    line_item_query.delete()
-
-        return True
 
     def _update_mappings(self):
         """Update cache of database objects for reference."""
