+++ koku/masu/test/processor/aws/test_aws_report_processor.py
@@ -173,12 +173,12 @@ class AWSReportProcessorTest(MasuTestCase):
             counts[table_name] = count
 
         bill_date = self.manifest.billing_period_start_datetime.date()
-        expected = f'INFO:masu.processor.aws.aws_report_processor:Deleting data for schema: acct10001 and bill date: {bill_date}'
+        expected = f'INFO:masu.processor.report_processor_base:Deleting data for schema: acct10001 and bill date: {bill_date}'
         logging.disable(
             logging.NOTSET
         )  # We are currently disabling all logging below CRITICAL in masu/__init__.py
         with self.assertLogs(
-            'masu.processor.aws.aws_report_processor', level='INFO'
+            'masu.processor.report_processor_base', level='INFO'
         ) as logger:
             processor.process()
             self.assertIn(expected, logger.output)
@@ -966,7 +966,7 @@ class AWSReportProcessorTest(MasuTestCase):
             manifest_id=self.manifest.id,
         )
         processor.process()
-        result = processor._delete_line_items()
+        result = processor._delete_line_items(AWSReportDBAccessor, self.column_map)
 
         with schema_context(self.schema):
             bills = self.accessor.get_cost_entry_bills()
@@ -988,7 +988,7 @@ class AWSReportProcessorTest(MasuTestCase):
             manifest_id=self.manifest.id,
         )
         processor.process()
-        result = processor._delete_line_items()
+        result = processor._delete_line_items(AWSReportDBAccessor, self.column_map)
         with schema_context(self.schema):
             bills = self.accessor.get_cost_entry_bills()
             for bill_id in bills.values():
@@ -1005,7 +1005,7 @@ class AWSReportProcessorTest(MasuTestCase):
             provider_id=self.aws_provider.id,
         )
         processor.process()
-        result = processor._delete_line_items()
+        result = processor._delete_line_items(AWSReportDBAccessor, self.column_map)
         with schema_context(self.schema):
             bills = self.accessor.get_cost_entry_bills()
             for bill_id in bills.values():
